backend/pyproject.toml
[project]
name = "thesisv2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiortc>=1.13.0",
    "anyio>=4.9.0",
    "fastapi[standard]>=0.116.1",
    "numpy>=2.3.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-ulid>=3.0.0",
    "scipy>=1.16.0",
    "scipy-stubs>=1.16.1.0",
    "streamz>=0.6.4",
    "vosk>=0.3.45",
    "websockets>=15.0.1",
]

[tool.pyrefly]
search-path = ["src/"]

[dependency-groups]
dev = [
    "pyrefly>=0.26.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.7",
]

[tool]
